+++ src/main/java/com/qa/sppd/card/Card.java
@@ -1,4 +1,4 @@
-package com.qa.main.card;
+package com.qa.sppd.card;
 
 import java.util.ArrayList;
 import java.util.Arrays;
