+++ src/test/java/com/qa/sppd/SppdCardsApplicationTests.java
@@ -1,4 +1,4 @@
-package com.qa.main;
+package com.qa.sppd;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
